+++ lib/helpers.js
@@ -12,6 +12,9 @@ exports.email = function(email, next) {
 exports.verify = function(assertion, audience, next) {
   verify(assertion, audience, config.get('verifier_url'),
     function(err, result) {
+      // In Hapi, the convention is to return a single result
+      // that may be a Hapi.Error or a valid result, instead
+      // of returning a separate error parameter ala node.js
       if (err) next(Hapi.Error.badRequest(err));
       else next(result.email);
     });
