+++ lib/prereqs.js
@@ -18,6 +18,15 @@ module.exports = {
     method: 'verify(payload.assertion, server.settings.uri)',
     assign: 'email'
   },
+  // This is the same the email prereq, but for GET requests
+  emailGet: {
+    method: 'email(query.email)'
+  },
+  // This is the same the assertion prereq, but for GET requests
+  assertionGet: {
+    method: 'verify(query.assertion, server.settings.uri)',
+    assign: 'email'
+  },
   userId: {
     method: 'userId(pre.email)'
   },
